refactor(web): lift enhance presets into a shared module
Extracts the preset catalog, preset card/picker, customize panel and the
params mapping out of MediaEnhancementDrawer into ./enhancePresets so the
upcoming batch dialog (issue #422) consumes exactly the same vocabulary
instead of a copy. Behaviour of the single-photo path is unchanged.

Relates to #422
